Реализован объектно-ориентированный улей и пасека.
Доработан механизм прохождения пчёлами поста охраны.

+++ VS Solution/Apiary/Apiary.BeeWorkflowApiary/Interfaces/IBeehive.cs
@@ -29,6 +29,6 @@ namespace Apiary.BeeWorkflowApiary.Interfaces
         /// Собрать накопленный в улье мёд.
         /// </summary>
         /// <returns>Количество собранного мёда.</returns>
-        int CollectHoney();
+        long CollectHoney();
     }
 }
